services/web_extractor/requirements.txt
requests
python-dotenv
bs4
openai
pydantic-settings
boto3
